requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.6
matplotlib==3.8.2
pandas==2.2.0
Pillow==10.2.0
PyYAML==6.0.1
jinja2==3.1.3
